Fix integer overflow in memory checkBounds (security audit 001)
Use long arithmetic in checkBounds to prevent signed integer overflow
when addr + size exceeds Integer.MAX_VALUE. Without this fix, the
bounds check silently passes and subsequent operations throw unhandled
NullPointerException or ArrayIndexOutOfBoundsException.

Affects both ByteArrayMemory and ByteBufferMemory.

+++ runtime/src/main/java/run/endive/runtime/ByteArrayMemory.java
@@ -320,7 +320,7 @@ public final class ByteArrayMemory implements Memory {
 
     private static void checkBounds(
             int addr, int size, int limit, Function<String, WasmEngineException> exceptionFactory) {
-        if (addr < 0 || size < 0 || addr > limit || (size > 0 && ((addr + size) > limit))) {
+        if (addr < 0 || size < 0 || addr > limit || (size > 0 && ((long) addr + (long) size > (long) limit))) {
             var errorMsg =
                     "out of bounds memory access: attempted to access address: "
                             + addr
